requirements.txt
torch
torchvision
facenet-pytorch
opencv-python
numpy
pandas
matplotlib
scikit-learn
tqdm
imutils
Pillow
